+++ lib/way_of_working/code_linting/hdi.rb
@@ -12,6 +12,7 @@ rescue LoadError # rubocop:disable Lint/SuppressedException
 end
 
 loader = Zeitwerk::Loader.for_gem_extension(WayOfWorking::CodeLinting)
+loader.collapse("#{__dir__}/hdi/generators/concerns")
 loader.ignore("#{__dir__}/hdi/plugin.rb")
 loader.setup