+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+<!-- markdownlint-disable-file MD024 -->
 # Changelog
 
 All notable changes to this project will be documented in this file.
@@ -6,6 +7,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Switched from Standard to ESLint (with Standard rules) to enable us to add Cypress and Jasmine globals
+- Added CODEOWNERS protection of the linting standards files
+
+### Changed
+
+- Switched to our custom HDI flavour of MegaLinter for faster GitHub action linting
+- Updated the linting GitHub workflow in line with upstream changes
 
 ## [1.0.0] - 2025-01-27